+++ CHANGES.rst
@@ -10,6 +10,9 @@ Changelog
 1.6rc2 (2018-03-22)
 -------------------
 
+- Ignore compressed data on auti log.
+  [cleberjsantos]
+
 - Fix profile version number.
   [hvelarde]